backend/requirements.txt
fastapi 
uvicorn 
requests 
python-dotenv 
pydantic 
pandas
openai 
httpx 
beautifulsoup4